CrossEncoder device (#2463)
* Move get_device_name method to util

* Add get_device_name from util, mps now available

* Add get_device_name method

* Change style with 'make style' to satisfy Quality checker

---------
